third_party/blink/web_tests/wpt_internal/webmidi/requestmidiaccess-basic.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
</head>
<body>
<script>
promise_test(async (t) => {
  assert_implements(navigator.requestMIDIAccess);
  await test_driver.set_permission({name: 'midi', sysex: true}, 'granted');
  const result = await navigator.requestMIDIAccess();
  assert_class_string(result, 'MIDIAccess');
}, 'requestMIDIAccess is defined and callable');
</script>
</body>
</html>
